// fe/src/app/api/admin/model/inventoryRequest.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface InventoryRequest { 
    id?: number;
    variantId?: number;
    stock?: number;
    importPrice?: number;
    sellingPrice?: number;
}