// fe/src/app/api/admin/model/importRequest.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ImportRequest { 
    variantId?: number;
    quantity?: number;
    importPrice?: number;
    note?: string;
}