// fe/src/app/api/admin/model/exportRequest.ts
/**
 * Shop Admin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface ExportRequest { 
    variantId?: number;
    quantity?: number;
    note?: string;
}